moved several functions from "run_loop" to another file and rewrote "show_attribute" function.
